Refactor refund handlers into separate files
