+++ apps/checkout-app/backend/payments/providers/adyen/refunds.ts
@@ -1,0 +1,6 @@
+import { TransactionRefund } from "@/checkout-app/types/refunds";
+
+export async function handleAdyenRefund(refund: TransactionRefund) {
+  await new Promise((resolve) => resolve(refund));
+  throw new Error("Not implemented");
+}
