+++ apps/checkout-app/backend/payments/providers/mollie/index.ts
@@ -1,2 +1,3 @@
 export * from "./createPayment";
 export * from "./webhookHandler";
+export * from "./refunds";
