+++ apps/checkout-app/types/refunds.ts
@@ -1,0 +1,5 @@
+export type TransactionRefund = {
+  id: string;
+  amount: string;
+  currency: string;
+};
